requirements.txt
setuptools>=69.0.3
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.20.0
pytest==8.3.3
pytest-django==4.7.0
pytest-asyncio==0.24.0
packaging==24.2
PyYAML==6.0.2
uritemplate==4.1.1
inflection==0.5.1
